requirements.txt
fastapi
uvicorn
streamlit
openai
faiss-cpu
sentence-transformers
neo4j
requests
